SideState should be a distinct abstract heap from Heap and Stack https://bugs.webkit.org/show_bug.cgi?id=145653
Reviewed by Geoffrey Garen.

Before, SideState fit into the hierarchy like so:

World
   |
   +-- Stack
   |
   +-- Heap
         |
         +-- SideState

Now we will have:

World
   |
   +-- Stack
   |
   +-- Heap
   |
   +-- SideState

This makes it easy to ask if a writing operation wrote to anything that is observable even
if we don't exit. SideState is only observable if we exit.

* dfg/DFGAbstractHeap.h:
(JSC::DFG::AbstractHeap::AbstractHeap):
(JSC::DFG::AbstractHeap::supertype):



Identifier: 163695@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185215 268f45cc-cd09-0410-ab3c-d52691b4dbfc