[iOS] Emoji overlap preceeding lines https://bugs.webkit.org/show_bug.cgi?id=145678 <rdar://problem/10684914>
Reviewed by Enrica Casucci.

Source/WebCore:

In WebKit, we first lay out lines, and then we lay out text inside those lines. Text is vertically centered in
its containing line. If the author has not specified a particular line-height, the height of a line comes from
the font metrics for the primary font of the line.

In iOS, we were specifically hardcoding the line height metric of Apple Color Emoji to be 0. This means that,
if Apple Color Emoji is the primary font for a line, and an explicit line-height was not specified, the lines
are laid out with 0 height. Then, when we vertically center the text inside the line, the top half of the text
overlaps the preceeding line.

I'm not sure exactly why were were hardcoding this value to 0, as it is surely wrong. I'm going to ask Enrica
to review this patch; hopefully she knows the answer.

Test: fast/text/emoji.html

* platform/graphics/cocoa/FontCocoa.mm:
(WebCore::Font::platformInit):

LayoutTests:

* fast/text/emoji.html:
* platform/mac/TestExpectations: Removed unnecessary line.
* platform/mac-mavericks/fast/text/emoji-expected.html:
* platform/mac/fast/text/emoji-expected.html:
* platform/ios-simulator/fast/text/emoji-expected.html:
* TestExpectations: Removed unnecessary line.

Identifier: 163745@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185266 268f45cc-cd09-0410-ab3c-d52691b4dbfc