[WK2] BrowsingContextController.loadData API is leaking Navigation objects https://bugs.webkit.org/show_bug.cgi?id=145977 <rdar://problem/21305706>
Reviewed by Darin Adler.

BrowsingContextController.loadData API was leaking Navigation objects.
The issue was due to WebPageProxy::loadData() creating a Navigation
object but not passing along the navigationID to the WebProcess when
sending the Messages::WebPage::LoadData IPC. As a result, the
WebProcess would use 0 as navigationID and we would end up creating a
new Navigation object later on for the same navigation.

* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::loadData):
* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::loadData):
* WebProcess/WebPage/WebPage.h:
* WebProcess/WebPage/WebPage.messages.in:


Identifier: 164025@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185563 268f45cc-cd09-0410-ab3c-d52691b4dbfc