AX: Changing state of radio buttons causes VoiceOver to go busy for a short time. https://bugs.webkit.org/show_bug.cgi?id=145933
Reviewed by Dean Jackson.

When radio buttons animate the new focus selection state, the thread activity looks a lot like short burst of
activity to draw, then wait on CoreAnimation to apply those changes.

Since those periods of activity during animation are so short, VoiceOver is not able to query for all the
attributes it needs, and gets stuck in the queue behind rendering.

The fix here is to turn off button state animations while VoiceOver is running.

* platform/mac/ThemeMac.mm:
(WebCore::updateStates):


Identifier: 164033@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185573 268f45cc-cd09-0410-ab3c-d52691b4dbfc