Played <audio> looks invisible against the gray background https://bugs.webkit.org/show_bug.cgi?id=146164 <rdar://problem/21014284>
Reviewed by Brent Fulgham.

The plus-darker blend mode was not allowing any white to
show through in the rendering. We don't need this for
audio controls, where we draw on an opaque grey background.

* Modules/mediacontrols/mediaControlsiOS.css:
(audio::-webkit-media-controls-panel): Darken the color of the controls a
little to make white stand out more.
(audio::-webkit-media-controls-timeline): Remove the plus-darker blending.
(video::-webkit-media-controls-timeline): Apply blending only to video.

Identifier: 164214@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185771 268f45cc-cd09-0410-ab3c-d52691b4dbfc