Crash under WebCore::PageConsoleClient::addMessage attempting to log insecure content message in ImageDocument https://bugs.webkit.org/show_bug.cgi?id=146096
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-06-19
Reviewed by Timothy Hatcher.

Was able to reproduce this using a user stylesheet with an http css font
on a pdf (ImageDocument) main document loaded over https. Was unable to
create a reliable test for this scenario.

* page/PageConsoleClient.cpp:
(WebCore::getParserLocationForConsoleMessage):
The scriptableDocumentParser could be null, such as in an ImageDocument.

Identifier: 164224@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185781 268f45cc-cd09-0410-ab3c-d52691b4dbfc