A/B testing results should be shown even if they were submitted to different platforms https://bugs.webkit.org/show_bug.cgi?id=146219
Reviewed by Andreas Kling.

Fetch A/B testing results regardless of the platform to which results are submitted
by providing the platform ID to which the results were submitted for each test group.

* public/api/test-groups.php:
(main): Include the platform id in the test groups.
* public/v2/analysis.js:
(App.TestGroup._fetchTestResults): Fetch results from the platform associated with the group.


Identifier: 164292@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185852 268f45cc-cd09-0410-ab3c-d52691b4dbfc